Move arrow msg to separate crate
