+++ myceliald/Makefile
@@ -23,7 +23,7 @@ dev:
 	RUST_LOG=$(RUST_LOG) cargo-watch -c -q \
 		-i myceliald_state.sqlite \
 		-i myceliald_state.sqlite-journal \
-		-s 'cargo run -- --config=config.example.toml'
+		-s 'cargo run -r -- --config=config.example.toml'
 
 .PHONY: vendor
 vendor:
