+++ pipe/arrow_msg/Cargo.toml
@@ -1,11 +1,10 @@
 [package]
-name = "snowflake"
+name = "arrow_msg"
 version = "0.1.0"
 edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-snowflake-api = "0.4.0"
-section = { version = "0.2", path = "../.." }
-tokio = { version = "1", features = ["full"] }
+section = { version = "0.2", path = "../section/" }
+arrow = "49"
